requirements.txt
django-extensions
django-multiforloop